avoid measuring compilation time
In Julia, the code is compiled in first call to every function, thus compilation time was being measured here, and might be significant. Running the two commands before the time-measure is more appropriate (or, alternatively, using `@btime` from `BenchmarkTools`, except I don't know if you are willing to include other packages not from the base language).